+++ Coding/ban_ve_xe_khach_web/src/main/java/com/cnwtt/ban_ve_xe_khach/repository/DanhGiaRepository.java
@@ -1,6 +1,5 @@
 package com.cnwtt.ban_ve_xe_khach.repository;
 
-import com.cnwtt.ban_ve_xe_khach.entity.ChuyenXe;
 import com.cnwtt.ban_ve_xe_khach.entity.DanhGia;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -8,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DanhGiaRepository extends JpaRepository<DanhGia, Integer> {
+public interface DanhGiaRepository extends JpaRepository<DanhGia, Long> {
     Page<DanhGia> findAll(Pageable pageable);
 
 }
